Added add/delete/edit and transactions view to portfolio page. Also added create and delete portfolio functionality.

+++ server/database.sql
@@ -19,8 +19,8 @@ CREATE TABLE assets(
     asset_id SERIAL,
     portfolio_id UUID NOT NULL,
     asset_coin_id VARCHAR(255) NOT NULL,
-    asset_amount NUMERIC NOT NULL,
-    initial_price NUMERIC NOT NULL,
+    asset_amount DOUBLE PRECISION NOT NULL,
+    initial_price DOUBLE PRECISION NOT NULL,
     PRIMARY KEY (asset_id),
     FOREIGN KEY (portfolio_id) REFERENCES portfolios(portfolio_id)
 );
